fix fixture on RewardPool and re-organize
